pyproject.toml
[project]
name = "datathon2026-challenge-harness"
version = "0.1.0"
description = "Starter FastAPI harness for the Datathon 2026 listings challenge."
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.39.0,<2.0.0",
    "fastapi>=0.115.0,<1.0.0",
    "httpx>=0.28.0,<1.0.0",
    "mcp>=1.25.0",
    "milvus-lite>=2.5.1",
    "open-clip-torch>=2.31.0,<3.0.0",
    "pillow>=10.4.0,<12.0.0",
    "pymilvus>=2.6.0,<3.0.0",
    "timm>=1.0.15,<2.0.0",
    "torch>=2.5.0,<3.0.0",
    "uvicorn[standard]>=0.32.0,<1.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0,<9.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app", "apps_sdk", "services"]

[tool.pytest.ini_options]
pythonpath = ["."]
